switched to context model

+++ singlefileeo.tests/bowling kata demo/demo.cs
@@ -19,12 +19,12 @@ namespace eventorientation.tests.bowling_kata_demo
             var gameId = "42";
             
             foreach(var pins in game)
-                sut.Roll(gameId, pins);
-            Assert.Equal(133, sut.Score(gameId));
+                sut.Roll(gameId, (uint)pins);
+            Assert.Equal(133, (int)sut.Score(gameId));
         }
         
 
-        void Roll_many(BowlingGame bg, string gameId, int n, int pins) =>
+        void Roll_many(BowlingGame bg, string gameId, int n, uint pins) =>
             Enumerable.Range(0, n).ToList().ForEach(_ => bg.Roll(gameId, pins));
     }
 }
